Move makeOpReturnCommitment() to SendFundsHandlingApi
