+++ core/src/main/scala/org/bitcoins/core/api/wallet/FundTransactionHandlingApi.scala
@@ -1,7 +1,7 @@
 package org.bitcoins.core.api.wallet
 
 import org.bitcoins.core.api.wallet.db.AccountDb
-import org.bitcoins.core.protocol.transaction.{Transaction, TransactionOutput}
+import org.bitcoins.core.protocol.transaction.TransactionOutput
 import org.bitcoins.core.wallet.builder.{
   FundRawTxHelper,
   ShufflingNonInteractiveFinalizer
@@ -29,28 +29,5 @@ trait FundTransactionHandlingApi {
       fromTagOpt: Option[AddressTag],
       markAsReserved: Boolean)
       : Future[FundRawTxHelper[ShufflingNonInteractiveFinalizer]]
-
-  def makeOpReturnCommitment(
-      message: String,
-      hashMessage: Boolean,
-      feeRate: FeeUnit,
-      fromAccount: AccountDb
-  ): Future[Transaction]
-
-  def makeOpReturnCommitment(
-      message: String,
-      hashMessage: Boolean,
-      feeRateOpt: Option[FeeUnit],
-      fromAccount: AccountDb): Future[Transaction]
-
-  def makeOpReturnCommitment(
-      message: String,
-      hashMessage: Boolean,
-      feeRate: FeeUnit): Future[Transaction]
-
-  def makeOpReturnCommitment(
-      message: String,
-      hashMessage: Boolean,
-      feeRateOpt: Option[FeeUnit]): Future[Transaction]
 
 }
