+++ wallet/src/main/scala/org/bitcoins/wallet/internal/FundTransactionHandling.scala
@@ -5,23 +5,16 @@ import org.bitcoins.core.api.feeprovider.FeeRateApi
 import org.bitcoins.core.api.keymanager.BIP39KeyManagerApi
 import org.bitcoins.core.api.wallet.*
 import org.bitcoins.core.api.wallet.db.AccountDb
-import org.bitcoins.core.currency.{CurrencyUnit, CurrencyUnits, Satoshis}
 import org.bitcoins.core.hd.HDAccount
 import org.bitcoins.core.policy.Policy
-import org.bitcoins.core.protocol.script.ScriptPubKey
 import org.bitcoins.core.protocol.transaction.*
-import org.bitcoins.core.script.constant.ScriptConstant
-import org.bitcoins.core.script.control.OP_RETURN
-import org.bitcoins.core.util.BitcoinScriptUtil
 import org.bitcoins.core.wallet.builder.*
 import org.bitcoins.core.wallet.fee.FeeUnit
 import org.bitcoins.core.wallet.utxo.*
-import org.bitcoins.crypto.CryptoUtil
 import org.bitcoins.db.SafeDatabase
 import org.bitcoins.wallet.config.WalletAppConfig
 import org.bitcoins.wallet.models.{SpendingInfoDAO, TransactionDAO}
 import org.bitcoins.wallet.WalletLogger
-import scodec.bits.ByteVector
 import slick.dbio.{DBIO, DBIOAction, Effect, NoStream}
 
 import scala.concurrent.Future
@@ -233,131 +226,6 @@ case class FundTransactionHandling(
       )
 
       fundTxHelper
-    }
-  }
-
-  override def makeOpReturnCommitment(
-      message: String,
-      hashMessage: Boolean,
-      feeRate: FeeUnit,
-      fromAccount: AccountDb
-  ): Future[Transaction] = {
-    val messageToUse = if (hashMessage) {
-      CryptoUtil.sha256(ByteVector(message.getBytes)).bytes
-    } else {
-      if (message.length > 80) {
-        throw new IllegalArgumentException(
-          s"Message cannot be greater than 80 characters, it should be hashed, got $message"
-        )
-      } else ByteVector(message.getBytes)
-    }
-
-    val asm = Seq(OP_RETURN) ++ BitcoinScriptUtil.calculatePushOp(
-      messageToUse
-    ) :+ ScriptConstant(messageToUse)
-
-    val scriptPubKey = ScriptPubKey(asm)
-
-    val output = TransactionOutput(Satoshis.zero, scriptPubKey)
-
-    for {
-      fundRawTxHelper <- fundRawTransactionInternal(
-        destinations = Vector(output),
-        feeRate = feeRate,
-        fromAccount = fromAccount,
-        coinSelectionAlgo = CoinSelectionAlgo.RandomSelection,
-        fromTagOpt = None,
-        markAsReserved = true
-      )
-      tx <- finishSend(
-        fundRawTxHelper,
-        CurrencyUnits.zero,
-        feeRate,
-        Vector.empty
-      )
-    } yield tx
-  }
-
-  override def makeOpReturnCommitment(
-      message: String,
-      hashMessage: Boolean,
-      feeRateOpt: Option[FeeUnit],
-      fromAccount: AccountDb): Future[Transaction] = {
-    for {
-      feeRate <- determineFeeRate(feeRateOpt)
-      tx <- makeOpReturnCommitment(message, hashMessage, feeRate, fromAccount)
-    } yield tx
-  }
-
-  override def makeOpReturnCommitment(
-      message: String,
-      hashMessage: Boolean,
-      feeRate: FeeUnit): Future[Transaction] = {
-    for {
-      account <- accountHandling.getDefaultAccount()
-      tx <- makeOpReturnCommitment(message, hashMessage, feeRate, account)
-    } yield tx
-  }
-
-  override def makeOpReturnCommitment(
-      message: String,
-      hashMessage: Boolean,
-      feeRateOpt: Option[FeeUnit]): Future[Transaction] = {
-    for {
-      feeRate <- determineFeeRate(feeRateOpt)
-      tx <- makeOpReturnCommitment(message, hashMessage, feeRate)
-    } yield tx
-  }
-
-  private def determineFeeRate(feeRateOpt: Option[FeeUnit]): Future[FeeUnit] =
-    feeRateOpt match {
-      case None =>
-        feeRateApi.getFeeRate()
-      case Some(feeRate) =>
-        Future.successful(feeRate)
-    }
-
-  /** Takes a [[RawTxBuilderWithFinalizer]] for a transaction to be sent, and
-    * completes it by: finalizing and signing the transaction, then correctly
-    * processing and logging it
-    */
-  private def finishSend[F <: RawTxFinalizer](
-      rawTxHelper: FundRawTxHelper[F],
-      sentAmount: CurrencyUnit,
-      feeRate: FeeUnit,
-      newTags: Vector[AddressTag]
-  ): Future[Transaction] = {
-    val signed = rawTxHelper.signedTx
-
-    val processedTxF = for {
-      ourOuts <- addressHandling.findOurOutputs(signed)
-      creditingAmount = rawTxHelper.scriptSigParams.foldLeft(
-        CurrencyUnits.zero
-      )(_ + _.amount)
-      _ <- transactionProcessing.processOurTransaction(
-        transaction = signed,
-        feeRate = feeRate,
-        inputAmount = creditingAmount,
-        sentAmount = sentAmount,
-        blockHashOpt = None,
-        newTags = newTags
-      )
-    } yield {
-      logger.debug(
-        s"Signed transaction=${signed.txIdBE.hex} with outputs=${signed.outputs.length}, inputs=${signed.inputs.length}"
-      )
-
-      logger.trace(s"Change output(s) for transaction=${signed.txIdBE.hex}")
-      ourOuts.foreach { out =>
-        logger.trace(s"    $out")
-      }
-      signed
-    }
-
-    processedTxF.recoverWith { case _ =>
-      // if something fails, we need to unreserve the utxos associated with this tx
-      // and then propogate the failed future upwards
-      utxoHandling.unmarkUTXOsAsReserved(signed).flatMap(_ => processedTxF)
     }
   }
 }
